Reference shape

Bcube.center — Method

center(::AbstractShape)

Center of the AbstractShape.

Implementation

Specialize for better performances

Bcube.entity — Method

entity(s::AbstractShape, ::Val{D}) where D

Return the geometrical Entity corresponding to the AbstractShape of a given degree D.

Remark : Returned entity must be consistent with the corresponding Lagrange function space.

Bcube.face_area — Method

face_area(::AbstractShape)

Return the length/area of the faces of a shape.

Bcube.face_shapes — Method

face_shapes(shape::AbstractShape, i)

Shape of i-th shape of the input shape.

Bcube.face_shapes — Method

face_shapes(::AbstractShape)

Return a tuple of the Shape of each face of the given (cell) Shape. For instance, a Triangle has three faces, all of them are Line.

Bcube.faces2nodes — Method

faces2nodes(shape::AbstractShape, side)

Return the index of the vertices on the iside-th face of a shape. If side is positive, the face is oriented preserving the cell normal. If side is negative, the face is returned with the opposite direction (i.e reverse node order).

Bcube.faces2nodes — Method

faces2nodes(::AbstractShape)

Return the index of the vertices on the faces of a shape.

Bcube.get_coords — Method

get_coords(shape::AbstractShape,i)

Return the coordinates of the ith shape vertices. i can be a tuple of indices, then the multiples vertices's coordinates are returned.

Bcube.get_coords — Method

get_coords(::AbstractShape)

Return node coordinates of the shape in the reference space.

Bcube.nedges — Method

nedges(::AbstractShape)

Generic function. Indicate how many edges a shape has.

Bcube.nfaces — Method

nfaces(::AbstractShape)

Indicate how many faces a shape has.

Bcube.normal — Method

normal(shape::AbstractShape, i)

Return the outward normal of the ith face of the shape.

Bcube.normals — Method

normals(::AbstractShape)

Return the outward normals of all the faces of the shape.

Bcube.nvertices — Method

nvertices(::AbstractShape)

Indicate how many vertices a shape has.

Bcube.shape — Method

shape(::AbstractEntityType)

Return the reference Shape corresponding to the given AbstractEntityType.