Function spaces

Bcube.AbstractFunctionSpaceType — Type

Abstract structure for the different types of function space, for instance the Lagrange function space, the Taylor function space etc.

Bcube.FunctionSpace — Method

FunctionSpace(fstype::Symbol, degree::Integer)
FunctionSpace(fstype::AbstractFunctionSpaceType, degree::Integer)

Build a FunctionSpace of the designated FunctionSpaceType and degree.

Examples

julia> FunctionSpace(:Lagrange, 2)
FunctionSpace{Bcube.Lagrange{:Uniform}, 2}()

Bcube.basis_functions_style — Method

basis_functions_style(fs::AbstractFunctionSpace)

Return the style (modal or nodal) corresponding to the basis functions of the 'fs'.

Bcube.coords — Method

coords(fs::AbstractFunctionSpace,::AbstractShape)

Return node coordinates in the reference space for associated function space and shape.

Bcube.get_degree — Method

get_degree(::AbstractFunctionSpace{type, degree}) where{type, degree}

Return the degree associated to the AbstractFunctionSpace.

Bcube.get_type — Method

get_type(::AbstractFunctionSpace{type})

Getter for the type of the AbstractFunctionSpace

Bcube.idof_by_edge — Method

idof_by_edge(::AbstractFunctionSpace, ::AbstractShape)

Return the local indices of the dofs lying on each edge of the Shape.

Dofs lying on the edge vertices are excluded.

The result is a Tuple of arrays of integers. Arrays maybe be empty. See Lagrange interpolation for simple examples.

Bcube.idof_by_edge_with_bounds — Method

idof_by_edge_with_bounds(::AbstractFunctionSpace, ::AbstractShape)

Return the local indices of the dofs lying on each edge of the Shape.

Dofs lying on the edge vertices are included.

The result is a Tuple of arrays of integers. Arrays maybe be empty. See Lagrange interpolation for simple examples.

Bcube.idof_by_face — Method

idof_by_face(::AbstractFunctionSpace, ::AbstractShape)

Return the local indices of the dofs lying on each face of the Shape.

Dofs lying on the face edges are excluded, only "face-interior" dofs are considered.

The result is a Tuple of arrays of integers. Arrays maybe be empty. See Lagrange interpolation for simple examples.

Bcube.idof_by_face_with_bounds — Method

idof_by_face_with_bounds(::AbstractFunctionSpace, ::AbstractShape)

Return the local indices of the dofs lying on each face of the Shape.

Dofs lying on the face edges are included

The result is a Tuple of arrays of integers. Arrays maybe be empty. See Lagrange interpolation for simple examples.

Bcube.idof_by_vertex — Method

idof_by_vertex(::AbstractFunctionSpace, ::AbstractShape)

Return the local indices of the dofs lying on each vertex of the Shape.

Beware that we are talking about the Shape, not the EntityType. So 'interior' vertices of the EntityType are not taken into account for instance. See Lagrange interpolation for simple examples.

Bcube.ndofs — Method

ndofs(fs::AbstractFunctionSpace, shape::AbstractShape)

Number of dofs associated to the given interpolation.

Bcube.shape_functions — Method

shape_functions(::AbstractFunctionSpace, ::Val{N}, shape::AbstractShape, ξ) where N
shape_functions(::AbstractFunctionSpace, shape::AbstractShape, ξ)

Return the list of shape functions corresponding to a FunctionSpace and a Shape. N is the size of the finite element space (default: N=1 if the argument is not provided).

The result is a vector of all the shape functions evaluated at position ξ, and not a tuple of the different shape functions. This choice is optimal for performance.

Note : λ = ξ -> shape_functions(fs, shape, ξ); λ(ξ)[i] is faster than λ =shape_functions(fs, shape); λ[i](ξ)

Implementation

Default version, should be overriden for each concrete FunctionSpace.

Bcube.shape_functions_vec — Method

shape_functions_vec(fs::AbstractFunctionSpace{T,D}, ::Val{N}, shape::AbstractShape, ξ) where {D,N}

Return all the shape functions of FunctionSpace on a Shape evaluated in ξ as a vector.

N is the the size (number of components) of the finite element space.

shape_functions_vec(fs::AbstractFunctionSpace{T,D}, n::Val{N}, shape::AbstractShape) where {T,D, N}

The shape functions are returned as a vector of functions.

Implementation

This is implementation is not always valid, but it is for Lagrange and Taylor spaces (the only two spaces available up to 20/01/23).

Bcube._var_on_bnd_nodes_discontinuous — Method

Apply the FEFunction on the nodes of the fdomain using the FunctionSpace representation for the cells.

Bcube._var_on_nodes_discontinuous — Method

Apply the FEFunction on the nodes of the mesh using the FunctionSpace representation for the cells.

Bcube.var_on_bnd_nodes_discontinuous — Function

var_on_bnd_nodes_discontinuous(f::AbstractFEFunction, fdomain::BoundaryFaceDomain, degree::Integer=max(1, get_degree(get_function_space(get_fespace(f)))))

Returns an array containing the values of f interpolated to new DoFs on fdomain. The DoFs locations on fdomain correspond to those of a discontinuous FESpace with a :Lagrange function space of selected degree.

Bcube.var_on_centers — Method

var_on_centers(f::SingleFEFunction, mesh::Mesh)

Interpolate solution on mesh vertices.

The result is a (ncells, ncomps) matrix if ncomps > 1, or a (ncells) vector otherwise.

Bcube.var_on_nodes_discontinuous — Function

var_on_nodes_discontinuous(f::AbstractFEFunction, mesh::Mesh, degree::Integer=max(1, get_degree(get_function_space(get_fespace(f)))))

Returns an array containing the values of f interpolated to new DoFs. The DoFs correspond to those of a discontinuous cell variable with a :Lagrange function space of selected degree.

Bcube.var_on_vertices — Method

var_on_vertices(f::AbstractFEFunction, mesh::Mesh)

Interpolate solution on mesh vertices.

The result is a (nnodes, ncomps) matrix.