Mapping

Mappings

Bcube.mapping — Method

mapping(nodes, ::AbstractEntityType, ξ)

Map the reference shape on the local shape.

Implementation

This function must be implemented for all shape.

Bcube.mapping — Method

mapping(nodes, ::Bar2_t, ξ)

Map the reference 2-nodes bar [-1,1] on the local bar:

$F(\xi) = \dfrac{x_r - x_l}{2} \xi + \dfrac{x_r + x_l}{2}$

Bcube.mapping — Method

mapping(nodes, ::Bar3_t, ξ)

Map the reference 3-nodes bar on the local bar (using Lagrange)

Bcube.mapping — Method

mapping(nodes, cshape::AbstractShape, ξ)

Returns the mapping of the an abstract shape (=ref element) to a target element defined by its nodes.

For instance, if cshape == Line, then the mapping is the same wether the input is the Shape or a Bar2_t. However if the cell is of type Bar3_t, it is still the Bar2_t mapping that is returned.

Bcube.mapping — Method

mapping(nodes, ::Hexa27_t, ξ)

Map the reference 8-nodes cube [-1,1] x [-1,1] x [-1,1] on the 27-hexa.

Bcube.mapping — Method

mapping(nodes, ::Penta6_t, ξ)

Map the reference 6-nodes prism [0,1] x [0,1] x [-1,1] on the 6-penta (prism).

Bcube.mapping — Method

mapping(nodes, ::Tri6_t, ξ)

Map the reference 6-nodes triangle [0,1] x [0,1] on the P2 curved-triangle.

$F(\xi) = \sum \lambda_i(\xi) x_i$ where $\lambda_i$ are the Lagrange P2 shape functions and $x_i$ are the local curved-triangle vertices' coordinates.

Bcube.mapping — Method

mapping(nodes, ::Hexa8_t, ξ)

Map the reference 8-nodes cube [-1,1] x [-1,1] x [-1,1] on the 8-hexa.

Bcube.mapping — Method

mapping(nodes, ::Quad4_t, ξ)

Map the reference 4-nodes square [-1,1] x [-1,1] on the 4-quadrilateral.

Bcube.mapping — Method

mapping(nodes, ::Quad9_t, ξ)

Map the reference 4-nodes square [-1,1] x [-1,1] on the P2 curved-quadrilateral.

$F(\xi) = \sum \lambda_i(\xi) x_i$ where $\lambda_i$ are the Lagrange P2 shape functions and $x_i$ are the local curved-quadrilateral vertices' coordinates.

Bcube.mapping — Method

mapping(nodes, ::Tri3_t, ξ)

Map the reference 3-nodes Triangle [0,1] x [0,1] on the local triangle.

\[F(\xi \\ \eta) = (1 - \xi - \eta) M_1 + x M_2 + y M_3\]

Bcube.mapping_det_jacobian — Method

mapping_det_jacobian(nodes, etype::AbstractEntityType, ξ)

Absolute value of the determinant of the mapping Jacobian matrix, expressed in the reference element.

Implementation

Default version using mapping_jacobian, but can be specified for each shape.

Bcube.mapping_det_jacobian — Method

mapping_det_jacobian(nodes, ::Bar2_t, ξ)

Absolute value of the determinant of the mapping Jacobian matrix for the reference 2-nodes bar [-1,1] to the local bar mapping.

$|det(J(\xi))| = \dfrac{|x_r - x_l|}{2}$

Bcube.mapping_det_jacobian — Method

mapping_det_jacobian(nodes, ::Tri3_t, ξ)

Absolute value of the determinant of the mapping Jacobian matrix for the the reference 3-nodes Triangle [0,1] x [0,1] to the local triangle mapping.

$|J| = |(x_2 - x_1) (y_3 - y_1) - (x_3 - x_1) (y_2 - y_1)|$

Bcube.mapping_det_jacobian — Method

mapping_det_jacobian(nodes, ::Quad4_t, ξ)

Absolute value of the determinant of the mapping Jacobian matrix for the the reference square [-1,1] x [-1,1] to the 4-quadrilateral mapping.

Bcube.mapping_face — Method

mapping_face(cshape::AbstractShape, side, permutation)

Build a mapping from the face reference element (corresponding to the side-th face of cshape) to the cell reference element (i.e the cshape), using a permutation of the face nodes.

Bcube.mapping_face — Method

mapping_face(cshape::AbstractShape, side)

Build a mapping from the face reference element (corresponding to the side-th face of cshape) to the cell reference element (i.e the cshape).

Implementation

We could define this function as an alias to mapping_face(cshape, side, 1:nnodes(face_shapes(cshape, side)) but for performance issue, I prefer to keep two independant functions for now.

Bcube.mapping_inv — Method

mapping_inv(nodes, ::AbstractEntityType, x)

Map the local shape on the reference shape.

Implementation

This function does not have to be implemented for all shape.

Bcube.mapping_inv — Method

mapping_inv(nodes, ::Bar2_t, x)

Map the local bar on the reference 2-nodes bar [-1,1]:

$F^{-1}(x) = \dfrac{2x - x_r - x_l}{x_r - x_l}$

Bcube.mapping_inv — Method

mapping_inv(nodes, ::Quad4_t, x)

Map the PARALLELOGRAM quadrilateral on the reference 4-nodes square [-1,1] x [-1,1]. Warning : this mapping is only corrects for parallelogram quadrilateral, not for any quadrilateral.

TODO: check this formulae with SYMPY

\[F^{-1} \begin{pmatrix} x \\ y \end{pmatrix} = \begin{pmatrix} a_1 x + b_1 y + c_1 \\ a_2 x + b_2 y + c_2 \end{pmatrix}\]

with

\[\begin{aligned} a_1 & = \dfrac{-2 (y_3-y_2)}{\Delta} \\ b_1 & = \dfrac{2 (x_3-x_2)}{\Delta} \\ c_1 & = -1 - a_1 x_1 - b_1 y_1 \\ a_2 & = \dfrac{-2 (y_1-y_2)}{\Delta} \\ b_2 & = \dfrac{2 (x_1 - x_2)}{\Delta} \\ c_2 & = -1 - a_2 x_1 - b_2 y_1 \end{aligned}\]

where $\Delta = (x_1 - x_2)(y_3 - y_2) - (x_3 - x_2)(y_1 - y_2)$

Bcube.mapping_inv — Method

mapping_inv(nodes, ::Tri3_t, x)

Map the local triangle on the reference 3-nodes Triangle [0,1] x [0,1].

TODO: check this formulae with SYMPY

\[F^{-1} \begin{pmatrix} x \\ y \end{pmatrix} = \frac{1}{x_1(y_2-y_3) + x_2(y_3-y_1) + x_3(y_1-y_2)} \begin{pmatrix} (y_3-y_1)x + (x_1 - x_3)y + (x_3 y_1 - x_1 y_3) \\ (y_1-y_2)x + (x_2 - x_1)x + (x_1 y_2 - x_2 y_1) \end{pmatrix}\]

Bcube.mapping_inv_jacobian — Method

mapping_inv_jacobian(nodes, etype::AbstractEntityType, x)

Jacobian matrix of the inverse mapping : $\dfrac{\partial F_i^{-1}}{\partial x_j}$

Implementation

Default version using LinearAlgebra to inverse the matrix, but can be specified for each shape (if it exists).

Bcube.mapping_inv_jacobian — Method

mapping_inv_jacobian(nodes, ::Bar2_t, x)

Mapping's jacobian matrix for the local bar to the reference 2-nodes bar [-1, 1].

$\dfrac{\partial F^{-1}}{\partial x} = \dfrac{2}{x_r - x_l}$

Bcube.mapping_inv_jacobian — Method

mapping_inv_jacobian(nodes, ::Tri3_t, x)

Mapping's jacobian matrix for the local triangle to the reference 3-nodes Triangle [0,1] x [0,1] mapping.

TODO: check this formulae with SYMPY

\[\frac{\partial F_i^{-1}}{\partial x_j} = \frac{1}{x_1 (y_2 - y_3) + x_2 (y_3 - y_1) + x_3 (y_1 - y_2)} \begin{pmatrix} y_3 - y_1 & x_1 - x_3 \\ y_1 - y_2 & x_2 - x_1 \end{pmatrix}\]

Bcube.mapping_jacobian — Method

mapping_jacobian(nodes, etype::AbstractEntityType, ξ)

Jacobian matrix of the mapping : $\dfrac{\partial F_i}{\partial \xi_j}$.

Implementation

Default version using ForwardDiff, but can be specified for each shape.

Bcube.mapping_jacobian — Method

mapping_jacobian(nodes, ::Bar2_t, ξ)

Mapping's jacobian matrix for the reference 2-nodes bar [-1, 1] to the local bar.

$\dfrac{\partial F}{\partial \xi} = \dfrac{x_r - x_l}{2}$

Bcube.mapping_jacobian — Method

mapping_jacobian(nodes, ::Bar3_t, ξ)

Mapping's jacobian matrix for the reference 2-nodes bar [-1, 1] to the local bar.

``\dfrac{\partial F}{\partial \xi} = \frac{1}{2} \left( (2\xi - 1) M1 + (2\xi + 1)M2 - 4 \xi M_3\right)

Bcube.mapping_jacobian — Method

mapping_jacobian(nodes, ::Quad4_t, ξ)

Mapping's jacobian matrix for the reference square [-1,1] x [-1,1] to the 4-quadrilateral

\[\frac{\partial F}{\partial \xi} = -M_1 + M_2 + M_3 - M_4 + \eta (M_1 - M_2 + M_3 - M_4) \frac{\partial F}{\partial \eta} = -M_1 - M_2 + M_3 + M_4 + \xi (M_1 - M_2 + M_3 - M_4)\]

Bcube.mapping_jacobian — Method

mapping_jacobian(nodes, ::Tri3_t, ξ)

Mapping's jacobian matrix for the reference 3-nodes Triangle [0,1] x [0,1] to the local triangle mapping.

\[\dfrac{\partial F_i}{\partial \xi_j} = \begin{pmatrix} M_2 - M_1 & M_3 - M_1 \end{pmatrix}\]

Bcube.mapping_jacobian_inv — Method

mapping_jacobian_inv(nodes, etype::AbstractEntityType, ξ)

Inverse of the mapping jacobian matrix. This is not exactly equivalent to the mapping_inv_jacobian since this function is evaluated in the reference element.

Implementation

Default version using ForwardDiff, but can be specified for each shape.

Bcube.mapping_jacobian_inv — Method

mapping_jacobian_inv(nodes, ::Bar2_t, ξ)

Inverse of mapping's jacobian matrix for the reference 2-nodes bar [-1, 1] to the local bar.

$\dfrac{\partial F}{\partial \xi}^{-1} = \dfrac{2}{x_r - x_l}$

Bcube.mapping_jacobian_inv — Method

mapping_jacobian_inv(nodes, ::Bar3_t, ξ)

Inverse of mapping's jacobian matrix for the reference 2-nodes bar [-1, 1] to the local bar.

$\dfrac{\partial F}{\partial \xi}^{-1} = \frac{2}{(2\xi - 1) M_1 + (2\xi + 1)M_2 - 4 \xi M_3}$

Bcube.mapping_jacobian_inv — Method

mapping_jacobian_inv(nodes, ::Tri3_t, ξ)

Inverse of mapping's jacobian matrix for the reference 3-nodes Triangle [0,1] x [0,1] to the local triangle mapping.

\[\dfrac{\partial F_i}{\partial \xi_j}^{-1} = \frac{1}{(x_1 - x_2)(y_1 - y_3) - (x_1 - x_3)(y_1 - y_2)} \begin{pmatrix} -y_1 + y_3 & x_1 - x_3 \\ y_1 - y_2 & -x_1 + x_2 \end{pmatrix}\]

Bcube.mapping_jacobian_inv — Method

mapping_jacobian(nodes, ::Quad4_t, ξ)

Inverse of mapping's jacobian matrix for the reference square [-1,1] x [-1,1] to the 4-quadrilateral

Reference to local